Substring-words and weights
added flag to track wether or not a node is a complete word, allowing for "substring-words" where one word is contained in another. it will now return both options where applicable.

Weights added to sort output words rather than alphabetically. Not fully implemented, still need to sort by the new weight.

function added to increment weight of node, should be used when a user selects one of the prompted options on the front end